+++ labbook_FE/app/templates/stat_product.js
@@ -1,17 +1,17 @@
-function status_prod( prod_id, prod_stat )
+function status_samp( samp_id, samp_stat )
 {
 let res = '' ;
 
-    if ( prod_stat == 9 )
-    res = '<button type="button" onclick="det_product(' + prod_id + ');" class="btn btn-danger">{{ _("A saisir") }}</span>' ;
-    else if ( prod_stat == 8 || prod_stat == 10 )
-    res = '<button type="button" onclick="det_product(' + prod_id + ');" class="btn btn-warning">{{ _("Modifier") }}</span>' ;
+    if ( samp_stat == 9 )
+    res = '<button type="button" onclick="det_sample(' + samp_id + ');" class="btn btn-danger">{{ _("A saisir") }}</span>' ;
+    else if ( samp_stat == 8 || samp_stat == 10 )
+    res = '<button type="button" onclick="det_sample(' + samp_id + ');" class="btn btn-warning">{{ _("Modifier") }}</span>' ;
 
 return res ;
 }
 
-function det_product(prod_id)
+function det_sample(samp_id)
 {
-window.location = "{{ session['server_ext'] }}/{{ session['redirect_name'] }}/det-product/" + prod_id ;
+window.location = "{{ session['server_ext'] }}/{{ session['redirect_name'] }}/det-sample/" + samp_id ;
 }