+++ CHANGELOG.md
@@ -3,11 +3,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [3.0.9] - 2021-06-21
+### Added
+- Configurable epidemio page
+- new sample analysis PB25
+
+### Fixed
+- rewritten PDF report function
+- groups analysis families on PDF report
+
 ## [3.0.8] - 2021-06-10
 ### Added
 - New analyzes B796, B797, B798, B799, B800, B801, B802, E06
 - DHIS2 filter by Sex and/or Age
-- DHIS2 filter on specific Record(s) 
+- DHIS2 filter on specific analysis 
 
 ### Changed
 - restore and backup not synchronous anymore